pyproject.toml
[project]
name = "specula"
version = "0.1.0"
description = "Speculative decoding research on Snapdragon X2 Elite Extreme"
requires-python = ">=3.12"
dependencies = [
    # HF model download & conversion (for future EAGLE-3 / DFlash GGUF conversion work)
    "huggingface-hub>=0.24",
    "transformers>=4.51",
    "safetensors>=0.4",
    "sentencepiece>=0.2",
    # Analysis & plotting
    "pandas>=2.2",
    "numpy>=2.0",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    # Notebook
    "jupyterlab>=4.2",
    # Prompt corpus handling
    "datasets>=3.0",
    # HTTP for result uploads / sharing
    "httpx>=0.27",
]

[project.optional-dependencies]
# Only needed if we want to run z-lab's reference DFlash (CUDA) on a different
# machine, or if we run Transformers-based inference locally on CPU for
# ground-truth comparisons against llama.cpp.
reference = [
    "torch>=2.5",
    "accelerate>=0.34",
]

# Phase 5 NPU drafting path. Pinned per docs/npu_scoping.md section 4.
# ORT-QNN bundles a signed QAIRT stack, which is the only path that loads on
# retail Windows on ARM without Secure Boot tricks.
npu = [
    "onnxruntime-qnn==1.24.4",
    "qai-hub>=0.48",
    "onnx>=1.17",
    # Hugging Face tooling for downloading pre-exported ONNX + BPE tokenizer.
    # NOTE: torch + optimum were initially in this list for a local
    # Qwen3 ONNX export, but torch has no cp312 win_arm64 wheel (only
    # cp38-cp311 on older 2.1.2). We bypass by using onnx-community's
    # pre-exported Qwen3-0.6B-ONNX instead. Local re-export would need
    # a non-WoA machine.
    "huggingface-hub>=0.24",
    "tokenizers>=0.20",
]

[tool.uv]
# Default to CPU torch on Windows ARM (no CUDA, no ROCm on WoA).
# Reference implementations run on a different machine; Windows-ARM Python is
# for tooling, not inference.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
